Convert session-value tests to use molecule
